+++ src/AppData/Doc/changelog.md
@@ -4,18 +4,6 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## v1.0.1.0 (2019-11-20)
-##### INFORMATION
-* Updated the project to C# 8.0
-
-## v1.0.1.0 (2019-11-14)
-##### ADDED
-* VerifyInpatientAdmissionDate.cs
-* CompleteOptionObject.cs
-##### CHANGED
-* AvatoolWebService.VerifyInpatientAdmissionDate() => VerifyInpatientAdmissionDate.VerifyInpatientAdmissionDate()
-* AvatoolWebService.CompleteOptionObject() => CompleteOptionObject.CompleteOptionObject()
-
 ## v1.0.0.0 (2019-11-14)
 ##### INFORMATION   
 * Initial release
